+++ joss/paper.md
@@ -31,10 +31,10 @@ Monte Carlo simulations are particularly effective at generating synthetic diffu
 `simDRIFT` was designed to be used by researchers of all disciplines and focuses who are working with diffusion MRI. Multiple scientific publications which utilize this library are currently in production. The wide customizability, high computational speed, and massively-parallel design will provide avenues for improved model development pipelines and thorough inter-model comparisons, among other potential applications. These same traits may also make `simDRIFT` useful for instructors of signal processing or diffusion imaging courses. 
 
 # Features
-Given the coarse-graining of detailed microstructural features (fiber bending, etc...) observed at experimentally realistic diffusion times and voxel sizes [@Novikov2018], `simDRIFT` represents fibers as narrow cylinders, or "sticks", and cells as isotropic spheres, or "balls" [@Behrens2003]. The library allows users to construct voxel geometries described by user-defined microstructural and scanning parameters. Specifically, `simDRIFT` simulates the diffusion MRI signal generated from the self-diffusion of water in an isotropic imaging voxel of length $L_{\mathrm{voxel}}$ that contains $n \in [0,4]$ distinct fiber bundles and $m \in [0,2]$ distinct cells types, according to the user's selection of diffusion imaging protocol, diffusion time ($\Delta$), time-step size $\mathrm{d}t$, and desired free water diffusivity $D_{FW}$. Within each simulated voxel, the user also has control over the properties of each fiber/cell type. For each fiber bundle, users define the desired orientation (via the angle $\theta$ formed between the bundle and the $z$ axis), intrinsic diffusivity ($D_{i}$), axonal radius ($R_{i}$), and voxel volume fraction $V_{i}/V_{\mathrm{vox}}$. For each cell type, users similarly define the desired radius $R_{j}$ and voxel volume fraction $V_{j}/V_{\mathrm{vox}}$. Examples of such voxel configurations can be seen in [Figure 1]. 
+Given the coarse-graining of detailed microstructural features (fiber bending, etc...) observed at experimentally realistic diffusion times and voxel sizes [@Novikov2018], `simDRIFT` represents fibers as narrow cylinders, or "sticks", and cells as isotropic spheres, or "balls" [@Behrens2003]. The library allows users to construct voxel geometries described by user-defined microstructural and scanning parameters. Specifically, `simDRIFT` simulates the diffusion MRI signal generated from the self-diffusion of water in an isotropic imaging voxel of length $L_{\mathrm{voxel}}$ that contains $n \in [0,4]$ distinct fiber bundles and $m \in [0,2]$ distinct cells types, according to the user's selection of diffusion imaging protocol, diffusion time ($\Delta$), time-step size $\mathrm{d}t$, and desired free water diffusivity $D_{FW}$. Within each simulated voxel, the user also has control over the properties of each fiber/cell type. For each fiber bundle, users define the desired orientation (via the angle $\theta$ formed between the bundle and the $z$ axis), intrinsic diffusivity ($D_{i}$), axonal radius ($R_{i}$), and voxel volume fraction $V_{i}/V_{\mathrm{vox}}$. For each cell type, users similarly define the desired radius $R_{j}$ and voxel volume fraction $V_{j}/V_{\mathrm{vox}}$. Examples of such voxel configurations can be seen in \ref{fig:examples}. 
 
-<p align = "center"> <img src = "https://github.com/jacobblum/simDRIFT/blob/main/joss/figs/joss_paper_figure.png" width = "650" height = "500"> </p> 
[…]
