### 2023-07-04 d7e7b6c
Update paper.md
Typo + figure caption fix
