+++ joss/paper.md
@@ -23,10 +23,10 @@ bibliography: paper.bib
 ---
 
 # Summary
-This library, `simDRIFT`, provides rapid and flexible Monte-Carlo simulations of diffusion-weighted magnetic resonance imaging (dMRI), which we expect to be useful for dMRI signal processing model development and validation purposes. The primary focus of this library is forward simulations of modular self-diffusion processes within an ensemble of nuclear magnetic resonance (NMR) active nuclei ("spins") residing in complex, biophysical tissue systems. To acheive a large variety of tissue configurations, `simDRIFT` provides support for $n$ fiber bundles (with user-defined radii, intrinsic diffusivities, orientation angles, and densities) and $m$ cells (with user-defined radii and volume fractions). simDrift is written in Python [Python Software Foundation, @VanRossum2010] and supported by a Numba [@Lam2015] backend. Thus, `simDRIFT` benefits from Numba's CUDA API, allowing the simulation of individual spin trajectories to be performed in parallel on single Graphics Processing Unit (GPU) threads. The resulting performance gains support `simDRIFT`'s aim to provide a customizable tool for the rapid prototyping of diffusion models, ground-truth model validation, and in silico phantom production.
+This library, `simDRIFT`, provides rapid and flexible Monte-Carlo simulations of diffusion-weighted magnetic resonance imaging (dMRI), which we expect to be useful for dMRI signal processing model development and validation purposes. The primary focus of this library is forward simulations of molecular self-diffusion processes within an ensemble of nuclear magnetic resonance (NMR) active nuclei ("spins") residing in complex, biophysical tissue systems. To acheive a large variety of tissue configurations, `simDRIFT` provides support for $n$ fiber bundles (with user-defined radii, intrinsic diffusivities, orientation angles, and densities) and $m$ cells (with user-defined radii and volume fractions). `simDrift` is written in Python [Python Software Foundation, @VanRossum2010] and supported by a Numba [@Lam2015] backend. Thus, `simDRIFT` benefits from Numba's CUDA API, allowing the simulation of individual spin trajectories to be performed in parallel on single Graphics Processing Unit (GPU) threads. The resulting performance gains support `simDRIFT`'s aim to provide a customizable tool for the rapid prototyping of diffusion models, ground-truth model validation, and in silico phantom production.
 
 # Statement of need
-Monte Carlo simulations are particularly effective at generating synthetic diffusion MRI data from complex, biophysically-accurate imaging voxels with known ground-truth microstructural parameters. Consequently, such simulations of the Brownian self-diffusion process have proven useful for developing and validating signal processing models [@Chiang2014; @Ye2020]. Existing Monte Carlo simulators typically rely on meshes to discretize the computational domain [see, e.g., @Panagiotaki2010, @Yeh2013, @Ianus2016, @Kerkelae2020, @RafaelPatino2020]. While this approach does allow for the representation of complex and finely-detailed microstructural elements, creating meshes for the biologically-relevant 3D geometries found in typical imaging voxels can be difficult and may therefore present a barrier to wide use among researchers who lack experience and training in computational mathematics. The software encompassed by `simDRIFT` therefore fulfills a presently-unmet need by allowing for mesh-free Monte Carlo simulations of dMRI that unify researchers' needs for computational performance and biophysical realism with easy-to-use and highly-configurable software.
[…]
