### 2023-07-24 #5621 @jacobblum
@djps

Thanks for trying it out. The main conda version the installation was tested on is 23.3.1 (on a windows 10 machine with nvcc 11.7.64). Trying things locally again, I don't see the same behavior, but this is very likely conda/cudatoolkit/nvcc version specific issue.

To try to narrow down what's happening could we run through a few details?

1. When the simDRIFT conda environment was created, was the python version was specified to be 3.8 instead of 3.11?
2. Is the error thrown as part of the cudatoolkit installation command or the PyTorch installation command?
3. Would you mind posting the error message or the associated error code?

Thanks so much!
