### 2023-07-31 #5621 @mfroeling
@kainenutt @jacobblum

Hi all I finally got around to evaluating the tool. It all seems to run but with some minor issues.

For me to check all the checklist items i think some things have to be addressed as listed below. Feel free to reach out if it needs more clarifications.

My major items are the documentation and statement of need / state of the field in the paper.

- [ ] when installing what aspects of the syntax for the GPU needs changing and how can i figure it out
- [ ] When running the example the matplotlib installation was not working. i reinstalled the matplotlib using conda and my environment was inconsistent, it was able to upgrade but it had to downgrade some packages (openssl cryptography and python). i would check the exported environement given in requirements.
- [ ] running the tests what is the expectation how long it should take or what its doing. the test ran for almost an hour on my laptop and what i could see only using the CPU. I have a CUDA enabled GPU put cannot test or see if it actually has been used. The test ran 20 items, i would like to see what is tested and maybe have an interface to test individual items.
- [ ] What system requirements are needed, can the tool be run on CPU only or does one need a GPU. how much memory does a typical simulation needs.

**examples**
although I'm not an python expert and hardly use it in the end i was able to run the examples
- [ ] . However i was a bit confused that the tool has to be called from the conda command window while the evaluation of the output is done in python. Could you give some examples of how to run the tool directly from python instead of the cmd. I can include os and use the command prompt using os.system but there must be a more elegant way to do this.
- [ ] In manual it is not clear that where you run the tool this is also the folder where the output is placed. When running the examples the use of directories is a bit messy. And only based on the python code I could figure out that some things like the b-values are packed in the toolbox while the output is done in the folder where the code is evaluated. But nowhere in the help i found what outputs to expect when running the code. Apparently the signals are stored in the signals folder as *.nii and the trajectories in a different folder. But nowhere is specified how these outputs should look like or what is in them.
- [ ] Can i use custom paths to the bval and bvec files
- [ ] The logfiles shows what has been performed but not the command used to generate these results so if i want to reproduce them and i didn't save the command I gave in the cmd window i have no clue how to reproduce them.
calling the tool
- [ ] It is not clear to me what the default values are and which inputs are mandatory. I removed some of the inputs of the examples and it still runs.
- [ ] Can the input settings move to a simulation config file where one can setup the simulation needed and just point to that file for all the inputs
- [ ] Is it possible to configure input and output directories for the simulation now things that were there are just overwritten so when running multiple simulations some management of input of output is needed or needs to be explained.
- [ ] from the manual it is not clear what the units of the inputs are. Diffusivity, bvalue, angles, time etc. etc. can all be specified in various methods, which units for each parameter does the tool assume

information needed
- [ ] which kind of diffusion weighting sequence is used (Stejskal Tanner?) does it take into account gradient slopes or does is assumes block gradients can the gradient wave forms be changed.
- [ ]  the bvalues and bvecs used should use the *.bval and *.bvec extensions in line with most diffusion toolboxes. how do these files need to be formatted.
- [ ] Is it possible to run the simulation with 3 fiber populations each along a different axes (i.e. x-axes, y-axes, z-axes) the examples given only show ration of the fiber population over one axes. In the brain crossing in 3 directions happens in many locations and given the example 2 i don't see how this tool can simulate this with only one rotation angle theta.

paper
- [ ] according to JOSS Substantial scholarly effort the software should "be a significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better". From the paper it is not fully clear to me what this tool can do different of other montecarlo simulation tools for diffusion such as CAMINO http://camino.cs.ucl.ac.uk/index.php?n=Tutorials.MCSimulator or difSim http://csci.ucsd.edu/software/difsim.html amongst others. The statement of need and state of the field are a bit non convincing for me. These kinds of simulations i have seen and used for years in various packages.

- [ ] The documentation does not contain a statement of need.
- [ ] Installation instructions are there, but they do not clearly list the needed dependencies in the toolbox. There is a cuda installation requirements.txt but i had to reinstall the matplotlib (my guess is due to the pytorch installation)
- [ ] Ideally the authors make a clean environment install only the minimum and do `conda env export | grep -v "^prefix: " > environment.yml` such that not each toolbox has to be installed manually as it is now.
- [ ] The documentation lack clear interface description (input and output definitions), which variables are needed which units, which files have to be stored or found where. how can the output be configured. What does the tool output (signals, trajectories) what are those file definitions what is stored in them and how can the be used.
- [ ] Testing is present but obscure, not sure what is tested. 20 tests are run but it is not displayed what is tested and why.
