+++ tests/test_suite.py
@@ -1,168 +1,0 @@
-import sys 
-import os 
-import numpy as np
-import pytest
-import nibabel as nb 
-from dipy.core.gradients import gradient_table
-import dipy.reconst.dti as dti
-import glob as glob
-from pathlib import Path
-from src.data import diffusion_schemes
-
-## Test Signals
-@pytest.mark.parametrize("expected", [('.nii',)])
-def test_signal_types(expected):
-    """1. Check that the forward simulated signal is a Nifti file
-    """
-    os.chdir(os.path.dirname(os.path.abspath(__file__)))
-    cmd = r"simDRIFT"
-    cmd += f" simulate --n_walkers 256000 --fiber_fractions 0,0 --cell_fractions 0,0 --Delta .001 --voxel_dims 10 --buffer 0 --verbose no" 
-    os.system(cmd)
-    signals = glob.glob(os.getcwd() + os.sep + 'signals' + os.sep + '*')
-    for signal in signals:
-        assert os.path.splitext(signal)[-1:] == expected
-
-
-@pytest.mark.parametrize("expected", [('.npy'),])
-def test_trajectory_types(expected):
-    """1. Check that the forward simulated trajectory is a .npy file
-    """
-    os.chdir(os.path.dirname(os.path.abspath(__file__)))
-    cmd = r"simDRIFT"
-    cmd += f" simulate --n_walkers 256000 --fiber_fractions 0,0 --cell_fractions 0,0 --Delta .001 --water_diffusivity {input} --voxel_dims 10 --buffer 0 --verbose no" 
-    os.system(cmd)
-    trajectories = glob.glob(os.getcwd() + os.sep + 'trajectories' + os.sep + '*')
-    for trajectory in trajectories:
-        assert os.path.splitext(trajectory)[-1:] == (expected,)
-
-@pytest.mark.parametrize("input, expected", [('DBSI_99', 99), ('ABCD', 103), ('NODDI_145', 145)])
-def test_signal_shapes(input, expected):
-    """1. Check that the forward simulated signal matches the number of bvals and bvecs used in the 'imaging' experiment 
-    """
-    os.chdir(os.path.dirname(os.path.abspath(__file__)))
-    cmd = r"simDRIFT"
-    cmd += f" simulate --n_walkers 256000 --fiber_fractions 0,0 --cell_fractions 0,0 --Delta 1 --diff_scheme {input} --voxel_dims 10 --buffer 0 --verbose no" 
-    os.system(cmd)
-    signal = nb.load(os.getcwd() + os.sep + 'signals' + os.sep + 'water_signal.nii').get_fdata()  
-    assert signal.shape == (expected,)
-
-@pytest.mark.parametrize("input, expected", [((os.path.join(Path(__file__).parents[1], 'src' + os.sep + 'data' + os.sep + 'bval99'), os.path.join(Path(__file__).parents[1], 'src' + os.sep + 'data' + os.sep + 'bvec99')), 99)])
-def test_custom_diffusion_scheme(input, expected):
-    """1. Check that the forward simulated signal matches the number of bvals and bvecs used in the PGSE experiment w/ 'custom' 
-       bvals and bvecs (i.e., loaded in from a path)
-    """
-    os.chdir(os.path.dirname(os.path.abspath(__file__)))
-    cmd = r"simDRIFT"
-    cmd += f" simulate --n_walkers 256000 --fiber_fractions 0,0 --cell_fractions .1,.1 --Delta 0.001 --voxel_dims 20 --buffer 0 --verbose no"
[…]
