+++ src/setup/objects.py
@@ -1,283 +1,0 @@
-import numpy as np
-
-
-
-class fiber():
-    """Class object for fiber attributes
-    """  
-    def __init__(self, center: np.ndarray, direction: np.ndarray, bundle: int, diffusivity: float, radius: float) -> None:
-        """Fiber information and parameters
-
-        :param center: (x,y,z) coordinates of the fiber center
-        :type center: np.ndarray
-        :param direction: Constituent bundle index
-        :type direction: np.ndarray
-        :param bundle: Unit vector pointed along the fiber direction
-        :type bundle: int
-        :param diffusivity: Intrinsic fiber diffusivity, square micrometers per millisecond
-        :type diffusivity: float
-        :param radius: Fiber radius, in micrometers
-        :type radius: float
-        """ 
-        self.center = center
-        self.bundle = bundle + 1
-        self.direction = direction
-        self.diffusivity = diffusivity
-        self.radius      = radius
-        return
-
-    """ Getters """
-        
-    def _get_center(self):
-        """Returns an array containing coordinates for the center of the specified fiber
-
-        :return: An array of fiber center coordinates
-        :rtype: numpy.ndarray
-        """  
-        return self.center
-    
-    def _get_bundle(self):
-        """Returns the fiber bundle index for the specified fiber
-
-        :return: The bundle/type index for the specified fiber
-        :rtype: int
-        """
-        return self.bundle
-    
-    def _get_direction(self):
-        """Returns the orientation vector for the specified fiber
-
-        :return: The orientation vector/rotation matrix for the specified fiber
-        :rtype: numpy.ndarray
-        """
-        return self.direction
-    
-    def _get_diffusivity(self):
-        """Returns the intra-axonal diffusivity for the specified fiber, in units of :math:`{\mathrm{μm}^2}\\, \mathrm{ms}^{-1}`
-
-        :return: The user-specified diffusivity within the specified fiber
-        :rtype: float
-        """
-        return self.diffusivity
-    
-    def _get_radius(self):
-        """Returns the radius of the specified fiber, in units of :math:`{\mathrm{μm}}`
-
-        :return: The user-specified radius of the specified fiber
-        :rtype: float
-        """
-        return self.radius
-    
-    
-
-
-
-class cell():
-    def __init__(self, cell_center, cell_radius: float, cell_diffusivity: float) -> None:
-        """Cell information
-
-        :param cell_center: (x,y,z) coordinates of the cell center 
-        :type cell_center: np.ndarray
-        :param cell_radius: Intrinsic fiber diffusivity
-        :type cell_radius: float
-        :param cell_diffusivity: Fiber radius
-        :type cell_diffusivity: float
-        """
-                
-        self.center      = cell_center
-        self.diffusivity = cell_diffusivity
-        self.radius      = cell_radius
-
-    """ Setters """     
-    def _set_center(self, cell_center: np.ndarray):
-        """Records the center coordinates for the specified cell
-
-        :param cell_center: An array containing the cell center coordinates
-        :type cell_center: np.ndarray
-        """        
-        self.center = cell_center
-        return
-    
-    def _set_diffusivity(self, D0: float):
-        """Records the diffusivity within the specified cell in units of :math:`{\mathrm{μm}^2}\\, \mathrm{ms}^{-1}`
-
-        :param D0: The diffusivity within the cell, assumed to be equal to the user-specified free water diffusivity
-        :type D0: float
-        """     
-        self.diffusivity = D0
-        return
-    
-    def _set_radius(self, radius: float):
-        """Records the radius of the specified cell, in units of :math:`{\mathrm{μm}}`
-
-        :param radius: The cell radius
-        :type radius: float
-        """        
-        self.radius = radius
-        return
-    
-    """ Getters """
-        
[…]
