+++ src/save.py
@@ -1,0 +1,212 @@
+import numpy as np  
+import os  
+import sys
+from src.data import diffusion_schemes
+import nibabel as nb
+import time
+import logging
+import torch 
+from datetime import datetime
+import shutil
+
+def _add_noise(signal, snr):
+    """Add Gaussian noise to the forward simulated signal [2]_
+
+    :param signal: The forward simulated signal
+    :type signal: np.ndarray
+    :param snr: The desired signal to noise ratio of the b0 image
+    :type snr: float
+    :return: The noise-added signal
+    :rtype: np.ndarray
+    """ 
+    sigma = 1.0/snr 
+    real_channel_noise = np.random.normal(0, sigma, signal.shape[0])
+    return signal + real_channel_noise
+
+   
+def _signal(spins: list, bvals: np.ndarray, bvecs: np.ndarray, Delta: float, dt: float, SNR = None) -> np.ndarray:
+    """Calculates the PGSE signal from the forward simulated spin trajectories [3]_. Note that this computation is executed on the GPU using PyTorch.
+
+    :param spins: A list of each objects.spin instance corresponding to a spin in the ensemble of random walkers
+    :type spins: list
+    :param bvals: The supplied b-values (diffusion weighting factors)
+    :type bvals: np.ndarray
+    :param bvecs: The supplied diffusion gradients 
+    :type bvecs: np.ndarray
+    :param Delta: The diffusion time, in milliseconds
+    :type Delta: float
+    :param dt: The time step parameter, also equal to delta because of the narrow pulse approximation
+    :type dt: float
+    :param SNR: The snr of the b0 image. If a value is not entered, the SNR of the signal is infinite. (Defaults to ``None``)
+    :type SNR: float, optional
+    :return: the forward simulated PGSE signal (``signal``), the initial spin positions (``trajectory_t1m``), and the final spin positions ``trajectory_t2p``
+    :rtype: np.ndarray
+    """
+    
+    gamma = 42.58 # MHz/T - The proton gyromagnetic ratio
+    delta = dt    # ms - From the narrow pulse approximation. More convenient for GPU memory management 
+
+    trajectory_t1m = torch.from_numpy(np.array([spin._get_position_t1m() for spin in spins])).float().to('cuda')
+    trajectory_t2p = torch.from_numpy(np.array([spin._get_position_t2p() for spin in spins])).float().to('cuda')
+    bvals_cu       = torch.from_numpy(np.array(bvals)).float().to('cuda')
+    bvecs_cu       = torch.from_numpy(np.array(bvecs)).float().to('cuda')
+
+    scaled_gradients = torch.einsum('i, ij -> ij', (torch.sqrt( (bvals_cu * 1e-3)/ (gamma**2*delta**2*(Delta - delta/3)))), bvecs_cu)
+    phase_shifts = gamma * torch.einsum('ij, kj -> ik', scaled_gradients, (trajectory_t1m - trajectory_t2p))*dt
+    signal =  torch.abs(torch.sum(torch.exp(-(0+1j)*phase_shifts), axis = 1)).cpu().numpy()
+    signal /= trajectory_t1m.shape[0]
+
+
+    if SNR != None:
+        noised_signal = _add_noise(signal, SNR)
+        return noised_signal, trajectory_t1m, trajectory_t2p
+    else:
+        return signal, trajectory_t1m, trajectory_t2p
+
+
+def _generate_signals_and_trajectories(self):
+    """Helper function to organize and store compartment specific and combined trajectories and their incident signals
+
+    :return: signals with associated labels (``signals_dict``), trajectories with associated labels (``trajectories_dict``)
+    :rtype: dictionaries
[…]
