+++ docs/source/usage-index.rst
@@ -1,0 +1,11 @@
+Usage
+=====
+
+remove-background
+-----------------
+
+Use case
+
+
+Recommended best practicies
+--------------------------
