### 2023-08-20 495642e
Update install-index.rst
added compatibility requirments
