+++ docs/source/usage-index.rst
@@ -1,11 +1,0 @@
-Usage
-=====
-
-remove-background
------------------
-
-Use case
-
-
-Recommended best practicies
---------------------------
