### 2023-08-21 e432981
Delete docs directory
