+++ docs/source/quickstart-index.rst
@@ -1,276 +1,0 @@
-*********************
-Quickstart
-*********************
-
-In this tutorial, we will show how to use ``simDRIFT`` to execute some example simulations. 
-Throughout, we will make use of a few Python libraries.
-
-.. code-block:: python
-    
-    import numpy as np     
-    import nibabel as nb 
-    import matplotlib.pyplot as plt 
-    import dipy.reconst.dti as dti
-    from dipy.core.gradients import gradient_table
-
-If you have an NVIDA-Cuda capable GPU, install ``simDRIFT`` by following these `instructions <https://simdrift.readthedocs.io/en/latest/install-index.html>`_ . If not, scroll to the bottom of the page and we will breifly go through an example using `Google Colab <https://colab.research.google.com/?utm_source=scs-index>`_. 
-
-Isotropic Diffusion
-----------------
-As a first example, let's simulate isotropic diffusion in an imaging voxel featuring only water. That is,
-our imaging voxel will feature no microstructural elements. To perform this simulation, nagivate to a directory of your choice
-and run ``simulate`` with the following command:
-
-.. code-block:: bash 
-    
-    (simDRIFT) >simDRIFT simulate --n_walkers 256000 --water_diffusivity 3.0 --fiber_fractions 0. --diff_scheme DBSI_99 --fiber_diffusions 0. --fiber_radii 1. --theta 0.
-
-The computation should finish within a minute or two, and the tool will produce the following files and directories:
-
-* ``trajectories`` : A directory under which .npy files corresponding to the by-compartment *(cells, fiber, water, etc...)* and total initial (*trajectories_t1m*) and final (*trajectories_t2p*)
-  spin positions are stored. The trajectory files may be useful for generating signals using a different diffusion scheme than the one provided 
-  by the ``diff_scheme`` argument post-hoc. 
-
-
-* ``signals`` : A directory under which NIfTI files containing the by-compartment and total signals generated from ``simDRIFT`` are stored. 
-
-* ``log`` : A text file that contains a detailed description of the input parameters and a record of the simulation's execution
-
-* ``input_configuration``: A copy of the input INI configuration file so that simulation input parameters may be referenced or simulations may be reproduced in the future. 
-
-
-
-We can analyze the results using a simple python script:
-
-.. code-block:: python
-
-    # Load Data
-    water_trajectories = np.load(r'PATH_TO_RESULTS/trajectories/water_trajectories_t2p.npy')
-
-    # Plot Results 
-    ax = plt.figure(figsize = (10,10)).add_subplot(projection = '3d')
-    ax.scatter(water_trajectories[1:,0], water_trajectories[1:,1], water_trajectories[1:,2], alpha = 0.05, color = 'purple', s = 1)
-
-    # For the figure legend
-    ax.scatter(water_trajectories[0,0], water_trajectories[0,1], water_trajectories[0,2], alpha = 0.5, color = 'purple', s = 50, label = 'water spin')
-    
-    ax.view_init(elev = 20., azim = 135)
[…]

[…]
