+++ docs/source/index.rst
@@ -1,0 +1,42 @@
+.. SimDRIFT documentation master file, created by
+   sphinx-quickstart on Wed Jun  7 12:43:30 2023.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+**************************************
+SimDRIFT Documentation Home
+**************************************
+
+*simDRIFT* is software package for massively parallel forward simulation of diffusion weighted MRI on biophysically accurate tissue systems
+
+For a brief survey of the package and its current capabilities, you can read
+the `introduction <https://numba.pydata.org/numba-doc/latest/user/installing.html>`_ 
+
+For details on input arguments and the output file format,
+see the `reference <https://numba.pydata.org/numba-doc/latest/user/installing.html>`_  section.
+
+For a step-by-step tutorial on running simulations with ``simDRIFT``, see the
+`quickstart <https://numba.pydata.org/numba-doc/latest/user/installing.html>`_  section.
+
+Contents
+----------------------
+.. toctree::
+   :maxdepth: 1
+
+   introduction/index
+   installation/index
+   reference/index
+   quickstart/index
+   simulation/index
+   theory/index
+  
+   
+Indices and tables
+----------------------
+
+* :ref:`genindex`
+* :ref:`search`
+
+.. note::
+   This project is under active development.
+    
