+++ docs/source/reference/index.rst
@@ -2,10 +2,10 @@ Reference
 =========
 
 simulate
------------------
+~~~~~~~~~~~
 
 Command Line Options
-~~~~~~~~~~~~~~~~~~~
+--------------------
 Typically, ``simDRIFT`` is run from the command line using the following command:
 
 .. code-block :: bash
@@ -13,7 +13,7 @@ Typically, ``simDRIFT`` is run from the command line using the following command
   (simDRIFT) >simDRIFT simulate --configuration CONFIGURATION.INI FILE
 
 Inputs
-~~~~~~~~~~~~~~~~~
+--------------------
 `simDRIFT` relies on a CONFIGURATION.INI file to set important simulation parameters. These parameters include:
 
 **Named Parameters**
@@ -38,7 +38,7 @@ Inputs
 - **water_diffusivity (float) -** The intrinsic diffusivity, in  :math:`μm^{2}/ms`, of the extra-fiber water. For simulations of in-vivo diffusion, we reccomend 3.0 :math:`μm^{2}/ms` and for ex-vivo simulations, :math:`2.0 μm^{2}/ms`.
 
 Example configuration file
-~~~~~~~~~~~~~~~~~~~~~~~~~~
+--------------------
 Below, please find an example of the structure of the CONFIGURATION.INI file used to run ``simDRIFT``.
 
 .. code-block:: bash
@@ -71,7 +71,7 @@ Below, please find an example of the structure of the CONFIGURATION.INI file use
 
 
 Outputs
-~~~~~~~~~~~~~~~~~
+--------------------
 Under the directory specified by the **output_directory** parameter, simDRIFT will create a directory titled ``DATE_TIME_simDRIFT_results``. Within this directory the tool will produce the following files and directories:
 
 * ``trajectories`` : A directory under which .npy files corresponding to the by-compartment *(cells, fiber, water, etc...)* and total initial (*trajectories_t1m*) and final (*trajectories_t2p*)
@@ -83,3 +83,87 @@ Under the directory specified by the **output_directory** parameter, simDRIFT wi
 * ``log`` : A text file that contains a detailed description of the input parameters and a record of the simulation's execution.
 
 * ``input_configuration``: A copy of the input INI configuration file so that simulation input parameters may be referenced or simulations may be reproduced in the future. 
+
+
+Editing the Configuration File Within a Python Script
+--------------------
+For the purposes of running batches of many number of simulations, an existing ``CONFIGURATION.INI`` file may easily be modified from within a Python script. Below, please 
+find an example code snippet used to modify a ``CONFIGURATION.INI`` used in the ``test suite``:
+
+.. code-block:: Python
+
+    import configparser
+
+    cfg_file = configparser.ConfigParser()
+    cfg_file.read(PATH_TO_CONFIG.INI FILE)
+
+    cfg_file['SIMULATION']['n_walkers'] = '256000'
+    cfg_file['SIMULATION']['DELTA'] = '.001'
+    cfg_file['SIMULATION']['dt'] = '.001'
+    cfg_file['SIMULATION']['voxel_dims'] = '10'
+    cfg_file['SIMULATION']['buffer'] = '0'
+    cfg_file['SIMULATION']['void_distance'] = '0'
+    cfg_file['SIMULATION']['bvals'] = "'N/A'"
+    cfg_file['SIMULATION']['bvecs'] = "'N/A'"
+    cfg_file['SIMULATION']['diffusion_scheme'] = "'DBSI_99'"
[…]
