+++ docs/source/index.rst
@@ -10,13 +10,13 @@ SimDRIFT Documentation Home
 *simDRIFT* is software package for massively parallel forward simulation of diffusion weighted MRI on biophysically accurate tissue systems
 
 For a brief survey of the package and its current capabilities, you can read
-the `introduction <https://numba.pydata.org/numba-doc/latest/user/installing.html>`_ 
+the `introduction <https://simdrift.readthedocs.io/en/latest/introduction/index.html>`_ 
 
-For details on input arguments and the output file format,
-see the `reference <https://numba.pydata.org/numba-doc/latest/user/installing.html>`_  section.
+For details on input arguments and the output file format, as well as how to edit and create the required configuration file, please
+see the `reference <https://simdrift.readthedocs.io/en/latest/reference/index.html>`_  section.
 
 For a step-by-step tutorial on running simulations with ``simDRIFT``, see the
-`quickstart <https://numba.pydata.org/numba-doc/latest/user/installing.html>`_  section.
+`quickstart <https://simdrift.readthedocs.io/en/latest/quickstart/index.html>`_  section.
 
 Contents
 ----------------------
