+++ src/config_empty_tempate.ini
@@ -1,0 +1,44 @@
+[SIMULATION]
+; The number of spins in the ensemble. Enter this value as an interger 
+n_walkers=
+; The diffusion time in the PGSE experiment (in ms)
+delta =
+; The timestep/gradient duration paramter (in ms)
+dt = 
+; The length of the isotropic imaging voxel (in um)
+voxel_dims = 75
+; The length (in um) beyond the voxel_dim over which microstructural elements will be placed. Useful for enforcing periodic boundary conditions.
+buffer = 0
+; The length (in um) between adjacent fiber bundles
+void_distance = 0
+; The absolute path to the bval file. See reference section of the documentation for a complete explanation of file formatting. 
+bvals = r'PATH_TO_BVAL_FILE'
+; The absolute path to the bvec file. See reference section of the documentation for a complete explanation of file formatting. 
+bvecs = r'PATH_TO_BVEC_FILE'
+; Keyword for one of the pre-implemented diffusion schemes. Choose one of 'DBSI_99', 'ABCD_103', or 'NODDI_145'
+diffusion_scheme = 'DBSI_99'
+; Absolute path to the directory under which outputs will be stored.
+output_directory = r'PATH_TO_OUTPUT_DIRECTORY'
+; Controlls logging level, enter 'no' to suppress logging outputs
+verbose = 'yes'
+
+[FIBERS]
+; Volume fraction of each fiber bundle, seperated by a comma.
+fiber_fractions = .2,.2,.2
+; Radius of each fiber in the fiber bundles specified above
+fiber_radii = 1.0,1.0,1.0
+; Angle (w.r.t the y-axis) in degrees of the fiber bundle
+thetas = 0,45,135
+; Intrinsic diffusivity (in um^2 / ms) of spins within the fiber
+fiber_diffusions = 1.0,2.0,3.0
+
+[CELLS]
+; Volume fraction of each cell population
+cell_fractions = 0
+; Radius of each cell in the above defined cell populations
+cell_radii = 1.0
+
+[WATER]
+; Intrinsic diffusivity of extra-fiber water (in um^2 / ms)
+water_diffusivity = 3.0
+
