+++ docs/source/reference/index.rst
@@ -16,26 +16,35 @@ Inputs
 --------------------
 `simDRIFT` relies on a CONFIGURATION.INI file to set important simulation parameters. These parameters include:
 
-**Named Parameters**
+**Simulation Parameters**
 
-- **n_walkers (int) -** The number of spins in the ensemble. We recommend using 256000 for quick, small scale simulations, and 1000000 for larger simulations. 
-- **Delta (float) -** The diffusion time (Δ), in milliseconds (ms), for the simulated PGSE experiment. 
-- **dt (float) -** The time step parameter, in milliseconds (ms), for the random walk. Because we adopt the narrow pulse approximation, this is also equal to the gradient duration (δ) used in the simulated PGSE experiment
-- **voxel_dims (float) -** The side-length, in micrometers (μm), of the isotropic imaging voxel. 
-- **buffer (float) -** The isotropic region, in micrometers (μm), beyond the imaging voxel where miscrostructural elements will still be populated. This parameter is useful for enforcing periodic boundary conditions.
-- **void_distance (float) -** The distance, in micrometers (μm), between adjacent fiber bundles. 
-- **bvals (str) -** The absolute path to a `.bval` file. The bval file should be in the FSL format, listing the b-value for each volume in the diffusion scheme. If you instead wish to use one of the already implemented diffusion schemes specified by the **diffusion_scheme** parameter, please enter 'N/A' here.
-- **bvecs (str) -** The absolute path to a `.bvec` file. The bvec file should be in the FSL format, listing the gradient direction, with one column per volume. Each column of the text file should have three rows, describing a unit vector with x,y,z components. If you instead wish to use one of the already implemented diffusion schemes specified by the **diffusion_scheme** parameter, please enter 'N/A' here.
-- **diffusion_scheme (str) -** We have already implemented three widely used diffusion schemes: The DBSI, ABCD, and NODDI schemes. To use one of these schemes, enter 'DBSI_99', 'ABCD_103', or 'NODDI_145' respectively. If paths are entered for the **bvecs** and **bvals** parameters, those inputs will take precedence over the input entered here, and the entry will be ignored.
-- **output_directory (str) -** The absolute path to the parent directory the simulation results will be stored under. If 'N/A' is entered, then the simulation will store results under the current working directory. 
-- **verbose (str) -** To suppress logging outputs, change the verbose parameter to 'no'. 
-- **fiber_fractions (float) -** The volume of the imaging voxel occupied by a fiber bundle. To place multiple fiber bundles, enter each fiber fraction seperated by a comma. For example, a single fiber bundle occupying 10 percent of the imaging voxel would be entered as **fiber_fractions** = 0.10, whereas three bundles each occupying 10 percent of the imaging voxel would be entered as **fiber_fractions** =0.10,0.10,0.10. Theoretically, there is no maximum number of fiber bundles allowed, however, we recomend not exceeding three or four.
[…]
