### 2023-08-21 35eccfa
Update config_empty_tempate.ini
