+++ src/config_empty_tempate.ini
@@ -11,14 +11,14 @@ voxel_dims = 75
 buffer = 0
 ; The length (in um) between adjacent fiber bundles
 void_distance = 0
-; The absolute path to the bval file. See reference section of the documentation for a complete explanation of file formatting. 
-bvals = r'PATH_TO_BVAL_FILE'
-; The absolute path to the bvec file. See reference section of the documentation for a complete explanation of file formatting. 
-bvecs = r'PATH_TO_BVEC_FILE'
-; Keyword for one of the pre-implemented diffusion schemes. Choose one of 'DBSI_99', 'ABCD_103', or 'NODDI_145'
+; The absolute path to the bval file. See reference section of the documentation for a complete explanation of file formatting. This entry MUST be preceeded by the "r" prefix to denote the string as literal. To use defaul diffusion scheme, enter 'N/A'.
+bvals = r'PATH_TO_BVAL_FILE' OR 'N/A'
+; The absolute path to the bvec file. See reference section of the documentation for a complete explanation of file formatting. This entry MUST be preceeded by the "r" prefix to denote the string as literal. To use defaul diffusion scheme, enter 'N/A'.
+bvecs = r'PATH_TO_BVEC_FILE' OR 'N/A'
+; Keyword for one of the pre-implemented diffusion schemes. Choose one of 'DBSI_99', 'ABCD_103', or 'NODDI_145'. If using this, enter 'N/A' for the bvals and bvecs 
 diffusion_scheme = 'DBSI_99'
-; Absolute path to the directory under which outputs will be stored.
-output_directory = r'PATH_TO_OUTPUT_DIRECTORY'
+; Absolute path to the directory under which outputs will be stored. To use the current working directory for the outputs, enter 'N/A'.
+output_directory = r'PATH_TO_OUTPUT_DIRECTORY' or 'N/A'
 ; Controlls logging level, enter 'no' to suppress logging outputs
 verbose = 'yes'
 
