### 2023-08-21 1e07d27
Update config_empty_tempate.ini
