+++ src/config_empty_tempate.ini
@@ -11,9 +11,9 @@ voxel_dims = 75
 buffer = 0
 ; The length (in um) between adjacent fiber bundles
 void_distance = 0
-; The absolute path to the bval file. See reference section of the documentation for a complete explanation of file formatting. This entry MUST be preceeded by the "r" prefix to denote the string as literal. To use defaul diffusion scheme, enter 'N/A'.
+; The absolute path to the bval file. See reference section of the documentation for a complete explanation of file formatting. This entry MUST be preceeded by the "r" prefix to denote the string as literal. To use a default diffusion scheme, enter 'N/A'.
 bvals = r'PATH_TO_BVAL_FILE' OR 'N/A'
-; The absolute path to the bvec file. See reference section of the documentation for a complete explanation of file formatting. This entry MUST be preceeded by the "r" prefix to denote the string as literal. To use defaul diffusion scheme, enter 'N/A'.
+; The absolute path to the bvec file. See reference section of the documentation for a complete explanation of file formatting. This entry MUST be preceeded by the "r" prefix to denote the string as literal. To use a default diffusion scheme, enter 'N/A'.
 bvecs = r'PATH_TO_BVEC_FILE' OR 'N/A'
 ; Keyword for one of the pre-implemented diffusion schemes. Choose one of 'DBSI_99', 'ABCD_103', or 'NODDI_145'. If using this, enter 'N/A' for the bvals and bvecs 
 diffusion_scheme = 'DBSI_99'
