+++ docs/source/installation/index.rst
@@ -17,7 +17,7 @@ First, create a conda environment and activate it:
    >conda create -n simDRIFT python=3.8
    >conda activate simDRIFT
 
-Then, install `numba <https://numba.pydata.org/numba-doc/latest/user/installing.html>`_  by following the linked instructions. For different hardware platforms, the specific numba installation syntax may varry. These instructions are covered in the `numba installation guide <https://numba.pydata.org/numba-doc/latest/user/installing.html>`_. Shown below are the commands for installation on our labs computers, which are x64-based windows machines.  
+Then, install `numba <https://numba.pydata.org/numba-doc/latest/user/installing.html>`_  by following the linked instructions. For different hardware platforms, the specific numba installation syntax may varry. These instructions are covered in the `numba installation guide <https://numba.pydata.org/numba-doc/latest/user/installing.html>`_. Shown below are the commands for installation on our lab's computers, which are x64-based windows machines.  
 
 .. code-block:: bash
    
