+++ docs/source/quickstart/index.rst
@@ -215,7 +215,7 @@ The signal can by analyzed with the below script
     print(1e3 * tenfit_1.ad, 1e3 * tenfit_2.ad, 1e3 * tenfit_3.ad, 1e3 * tenfit_water.ad, 1e3 * tenfit_water.rd)
 
 For the fibers, are estimated axial diffusivities are :math:`\lambda_{||}^{(1)} = 0.996 \mu m^{2} / ms`, :math:`\lambda_{||}^{(2)} = 2.007 \mu m^{2} / ms`,
-:math:`\lambda_{||}^{(3)} = 2.996 \mu m^{2} / ms`, and for the water, we get that: :math:`\lambda_{||} = 2.82 \mu m^{2} / ms` and :math:`\lambda_{\perp} = 2.73 \mu m^{2} / ms`. The fiber values are exactly in the range that we would expect. Of course, although the water diffusivity is set to :math:`3.0 :math:`, because of the diffusion restricting barriers imposed by the fiber bundles, we can no longer hope to recover this number exactly (at reasonably high fiber densities).
+:math:`\lambda_{||}^{(3)} = 2.996 \mu m^{2} / ms`, and for the water, we get that: :math:`\lambda_{||} = 2.82 \mu m^{2} / ms` and :math:`\lambda_{\perp} = 2.73 \mu m^{2} / ms`. The fiber values are exactly in the range that we would expect. Of course, although the water diffusivity is set to :math:`3.0 \mu m^{2} / ms`, because of the diffusion restricting barriers imposed by the fiber bundles, we can no longer hope to recover this number exactly (at reasonably high fiber densities).
 
 Google Colab
 --------------
