### 2023-08-21 #5621 @jacobblum
> @jacobblum - I can't seem to install it when following the instructions:
>
> In windows 11, using conda 23.5.2, (with python 3.11), cuda driver (528.90) and nvcc 11.3.58 but can't install the cuda-toolkit.
>
> It finds lots of conflicts which it can't resolve. What version of conda was this tested on? I had tried different channels also.
>
> I was able to install it using pip though. The tests passed after 28 minutes. I could replicate the quick start examples as well.

@djps, The error message you attached below precipitated by the conda install command indicates some sort of incompatibility with the llvmlite package, which binds Python to LLVM. The llvmlite readthedocs indicates that there are many potential sources of incompatibility between a given build of llvmlite and the LLVM release to which it is statically linked, which may be dependent on a machines architecture, among many other reasons (https://llvmlite.pydata.org/en/latest/faqs.html).

Unfortunately if this is the issue, I don't think there is much we can do about it in terms of completely resolving these incompatibilities among the dependencies. Luckily, however, given pip's less strict dependency verification processes (relative to conda), it appears that whatever incompatibility conda is finding, it is subtle and ultimately inconsequential given that the test suite still runs correctly when installation is facilitated by pip.

We have thus amended our installation (https://simdrift.readthedocs.io/en/latest/installation/index.html) documentation to make note of the potential architecture dependence, and specifically have deferred to Numba's installation guide, which covers the use of both pip and conda for installation purposes. This should help to avoid any hiccups with this step of the installation.
