### 2023-08-21 #5621 @jacobblum
>
>
> There are contributing guidelines, but no mention of them nor reporting issue or seeking support in the README
>
>

We have now added these elements to the README. Thanks!
