### 2023-08-24 7a4839a
Update index.rst
Consistent references to simDRIFT
