### 2023-08-24 faed6b8
Update index.rst
edits from sk song
