### 2023-08-24 2604c7f
Update index.rst
edits from sks
