### 2023-08-24 4f72afa
Update index.rst
Edits from SK song
