+++ docs/source/reference/index.rst
@@ -14,37 +14,37 @@ Typically, ``simDRIFT`` is run from the command line using the following command
 
 Inputs
 --------------------
-`simDRIFT` relies on a CONFIGURATION.INI file to set important simulation parameters. These parameters include:
+``simDRIFT`` relies on a CONFIGURATION.INI file to set important simulation parameters. These parameters include:
 
 **Simulation Parameters**
 
-- ``n_walkers (int) -`` The number of spins in the ensemble. We recommend using 256000 for quick, small scale simulations, and 1000000 for larger simulations. 
+- ``n_walkers (int) -`` The number of spins in the ensemble. We recommend using 256000 for quick, small-scale simulations, and 1000000 for larger simulations. 
 - ``Delta (float) -`` The diffusion time (Δ), in milliseconds (ms), for the simulated PGSE experiment. 
-- ``dt (float) -`` The time step parameter, in milliseconds (ms), for the random walk. Because we adopt the narrow pulse approximation, this is also equal to the gradient duration (δ) used in the simulated PGSE experiment
+- ``dt (float) -`` The time step parameter, in milliseconds (ms), for the random walk. Because we adopt the narrow pulse approximation, this is also equal to the gradient duration (δ) used in the simulated PGSE experiment.
 - ``voxel_dims (float) -`` The side-length, in micrometers (μm), of the isotropic imaging voxel. 
-- ``buffer (float) -`` The isotropic region, in micrometers (μm), beyond the imaging voxel where miscrostructural elements will still be populated. This parameter is useful for enforcing periodic boundary conditions.
+- ``buffer (float) -`` The isotropic region, in micrometers (μm), beyond the imaging voxel where microstructural elements will still be populated. This parameter is useful for enforcing periodic boundary conditions.
 - ``void_distance (float) -`` The distance, in micrometers (μm), between adjacent fiber bundles. 
-- ``bvals (str) -`` The absolute path to a `.bval` file. The bval file should be in the FSL format, listing the b-value for each volume in the diffusion scheme. If you instead wish to use one of the already implemented diffusion schemes specified by the ``diffusion_scheme`` parameter, please enter 'N/A' here. Importantly, the absolute path specified here **MUST be preceeded by the "r" prefix to denote the string as literal!**
-- ``bvecs (str) -`` The absolute path to a `.bvec` file. The bvec file should be in the FSL format, listing the gradient direction, with one column per volume. Each column of the text file should have three rows, describing a unit vector with x,y,z components. If you instead wish to use one of the already implemented diffusion schemes specified by the ``diffusion_scheme`` parameter, please enter 'N/A' here. Importantly, the absolute path specified here  **MUST be preceeded by the "r" prefix to denote the string as literal!**
+- ``bvals (str) -`` The absolute path to a `.bval` file. The bval file should be in the FSL format, listing the b-value for each volume in the diffusion scheme. If you instead wish to use one of the already implemented diffusion schemes specified by the ``diffusion_scheme`` parameter, please enter 'N/A' here. Importantly, the absolute path specified here **MUST be preceded by the "r" prefix to denote the string as literal!**
[…]
