### 2023-08-24 1b59151
Update index.rst
typo
