### 2023-08-24 #5621 @jacobblum
@mfroeling

Thank you for the very useful feedback, and for your patience while we addressed your comments. Please see our responses below. You will find that we have made many of the changes and clarifications you have suggested.

> @kainenutt @jacobblum
>
> Hi all I finally got around to evaluating the tool. It all seems to run but with some minor issues.
>
> For me to check all the checklist items i think some things have to be addressed as listed below. Feel free to reach out if it needs more clarifications.
>
> My major items are the documentation and statement of need / state of the field in the paper.
>
> **installation**
>
> * [ ]  when installing what aspects of the syntax for the GPU needs changing and how can i figure it out

Thank you for your question. While I’m not sure I’m understanding the question exactly, I want to the simulation will only run on the GPU. Without an available Cuda device, an error message would be thrown after the command line call preventing the simulation from running. See main/src/cli.py, lines 110-111 for details.

> * [ ]  When running the example the matplotlib installation was not working. i reinstalled the matplotlib using conda and my environment was inconsistent, it was able to upgrade but it had to downgrade some packages (openssl cryptography and python). i would check the exported environement given in requirements.

Thank you for pointing this out! We have now included a compatible matplotlib version in the setup.py requirements. It will automatically install while setting up simDRIFT now.

> * [ ]  running the tests what is the expectation how long it should take or what its doing. the test ran for almost an hour on my laptop and what i could see only using the CPU. I have a CUDA enabled GPU put cannot test or see if it actually has been used. The test ran 20 items, i would like to see what is tested and maybe have an interface to test individual items.

 Thank you for the comment here. We have included all test items in the documentation (https://simdrift.readthedocs.io/en/latest/test_suite/index.html) and added logging to the test suite so that it is clear what exactly is being tested. It is difficult to predict a runtime given the hardware dependence here; however, in the documentation we made note of the runtime on our hardware (about 8 minutes). We also have remarked that due to the small-scale of the test suite’s simulations, GPU utilization may not increase significantly even while the code is executing on the GPU. Like I mentioned in the above reply, the simulation cannot run without detecting and using the CUDA device.

> * [ ]  What system requirements are needed, can the tool be run on CPU only or does one need a GPU. how much memory does a typical simulation needs.

Thank you for the suggestion. The authors agree that this is great information to include, and so we have added this to the installation documentation (https://simdrift.readthedocs.io/en/latest/installation/index.html). Specifically, from a hardware standpoint we require a CUDA device with compute capability of 3.0 or higher. We have also made note that typical simulations do not exceed 2 Gb of VRAM use.

>
> **examples** although I'm not an python expert and hardly use it in the end i was able to run the examples
>
> * [ ]  . However i was a bit confused that the tool has to be called from the conda command window while the evaluation of the output is done in python. Could you give some examples of how to run the tool directly from python instead of the cmd. I can include os and use the command prompt using os.system but there must be a more elegant way to do this.

Thank you for your suggestion. While the authors understand the point made here, we believe that it is easiest to use the tool from the command line. As you mention, calling the program from within a Python script is easily done using os.system. This is a common practice within our research group’s workflow in the context of simulation and using other tools like FSL from within a Python script.

We recommend using a conda environment to prevent conflict with any other Python packages you may have installed for use in other workflows. Conda is just a package manager here, all the work, as you mention, is done with Python.

> * [ ]  In manual it is not clear that where you run the tool this is also the folder where the output is placed. When running the examples the use of directories is a bit messy. And only based on the python code I could figure out that some things like the b-values are packed in the toolbox while the output is done in the folder where the code is evaluated. But nowhere in the help i found what outputs to expect when running the code. Apparently the signals are stored in the signals folder as *.nii and the trajectories in a different folder. But nowhere is specified how these outputs should look like or what is in them.

Thank you for the useful comment here. We have added a reference section to the documentation that specifically addresses program inputs and outputs (https://simdrift.readthedocs.io/en/latest/reference/index.html)

> * [ ]  Can i use custom paths to the bval and bvec files

Yes, instructions regarding formatting and inputting custom .bval and .bvec files are now covered in the reference section of the documentation: (https://simdrift.readthedocs.io/en/latest/reference/index.html)

> * [ ]  The logfiles shows what has been performed but not the command used to generate these results so if i want to reproduce them and i didn't save the command I gave in the cmd window i have no clue how to reproduce them.

This is a great point. We have included in the outputs a copy of the input arguments, as now noted in the documentation (https://simdrift.readthedocs.io/en/latest/reference/index.html)

>
> **calling the tool**
>
> * [ ]  It is not clear to me what the default values are and which inputs are mandatory. I removed some of the inputs of the examples and it still runs.

We had previously implemented default inputs when the named parameter was omitted in the command line call. We have removed this and made note of all inputs in the reference section of the documentation (https://simdrift.readthedocs.io/en/latest/reference/index.html).

> * [ ]  Can the input settings move to a simulation config file where one can setup the simulation needed and just point to that file for all the inputs

This is a great suggestion, and we have implemented this functionality. As is covered in the reference section of the documentation, the simulation takes as input a INI configuration file that contains all important input parameters. Please see both the reference and quickstart section of the documentation for details.

> * [ ]  Is it possible to configure input and output directories for the simulation now things that were there are just overwritten so when running multiple simulations some management of input of output is needed or needs to be explained.

Thank you for the suggestion. This is a great idea, and the feature has now been implemented. Please see the reference or quickstart section of the documentation for details.

> * [ ]  from the manual it is not clear what the units of the inputs are. Diffusivity, bvalue, angles, time etc. etc. can all be specified in various methods, which units for each parameter does the tool assume

Thank you for the comment. All units have now been added to input parameter descriptions. Please see details in the reference section of the documentation (https://simdrift.readthedocs.io/en/latest/reference/index.html).

>
> **information needed**
>
> * [ ]  which kind of diffusion weighting sequence is used (Stejskal Tanner?) does it take into account gradient slopes or does is assumes block gradients can the gradient wave forms be changed.

Thank you for the question. Throughout the documentation we now make note that we have implemented the PGSE (Stejskal Tanner) experiment with the narrow pulse approximation.

> * [ ]   the bvalues and bvecs used should use the *.bval and *.bvec extensions in line with most diffusion toolboxes. how do these files need to be formatted.

Thank you for the question. The authors have clarified .bval and .bvec file formatting in the reference section of the documentation.

> * [ ]  Is it possible to run the simulation with 3 fiber populations each along a different axes (i.e. x-axes, y-axes, z-axes) the examples given only show rotation of the fiber population over one axes. In the brain crossing in 3 directions happens in many locations and given the example 2 i don't see how this tool can simulate this with only one rotation angle theta.

While the authors understand the point made here, we wanted to avoid populating the simulated image voxel with overlapping fibers, which is achieved by allowing rotation along more than one axis. The major point of emphasis in our simulation was to test DWI inverse problem model’s ability to resolve crossing fiber angles. To an extent, we view all crossing angles as isomorphic to the geometry implemented in the simulation. Crossing angles along several axes is therefore beyond the scope of this simulation.

>
> **paper**
>
> * [ ]  according to JOSS Substantial scholarly effort the software should "be a significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better". From the paper it is not fully clear to me what this tool can do different of other montecarlo simulation tools for diffusion such as CAMINO http://camino.cs.ucl.ac.uk/index.php?n=Tutorials.MCSimulator or difSim http://csci.ucsd.edu/software/difsim.html amongst others. The statement of need and state of the field are a bit non convincing for me. These kinds of simulations i have seen and used for years in various packages.

Thank you for the comment. A major point of emphasis in the simDRIFT statement of need is that this simulator, unlike others, is mesh free, allowing users to easily specify many potential image voxel configurations. By contrast, simulators like CAMINO rely on either already packaged or pre-supplied meshes for simulation, which can be difficult to make for complex voxel geometries. Therefore, simDRIFT allows users who lack experience or training in generating meshes for computational domains to simulate complex image voxel configurations.  Another important difference is that this simulation runs on the GPU, while CAMINO runs single threaded on the CPU.

>
> **Documentation**
>
> * [ ]  The documentation does not contain a statement of need.

We have included the statement of need in the documentation’s introduction section (https://simdrift.readthedocs.io/en/latest/introduction/index.html)

> * [ ]  Installation instructions are there, but they do not clearly list the needed dependencies in the toolbox. There is a conda/pip installation requirements.txt but i had to reinstall the matplotlib (my guess is due to the pytorch installation)

All dependencies should now be accounted for in the setup.py file.

> * [ ]  Ideally the authors make a clean environment install only the minimum and do `conda env export | grep -v "^prefix: " > environment.yml` such that not each toolbox has to be installed manually as it is now.

Thank you for your suggestion. The authors prefer the current installation instructions because they prompt the user to install appropriate NVIDIA drivers and hardware architecture specific versions of both Numba and Pytorch, without which the code would not run.

> * [ ]  The documentation lack clear interface description (input and output definitions), which variables are needed which units, which files have to be stored or found where. how can the output be configured. What does the tool output (signals, trajectories) what are those file definitions what is stored in them and how can the be used.

All this information should now be included in the reference section of the documentation, as noted above.

> * [ ]  Testing is present but obscure, not sure what is tested. 20 tests are run but it is not displayed what is tested and why.

Thank you for the comment here. We have included all test items in the documentation (https://simdrift.readthedocs.io/en/latest/test_suite/index.html) and added logging to the test suite so that it is clear what exactly is being tested, as described above.
