+++ examples/qs_isotropic_config.ini
@@ -1,0 +1,26 @@
+[SIMULATION]
+n_walkers = 256000
+delta = 10
+dt = .001
+voxel_dims = 75
+buffer = 0
+void_distance = 0
+bvals = 'N/A'
+bvecs = 'N/A'
+diffusion_scheme = 'DBSI_99'
+output_directory = 'N/A'
+verbose = 'yes'
+
+[FIBERS]
+fiber_fractions = 0,0
+fiber_radii = 1.0,1.0
+thetas = 0,0
+fiber_diffusions = 1.0,2.0
+
+[CELLS]
+cell_fractions = 0
+cell_radii = 1.0
+
+[WATER]
+water_diffusivity = 3.0
+
