### 2023-09-05 #5621 @mfroeling
@jacobblum

The clarifications and updates realy helped. I reinstalled everything and was better able to understand the steps and examples since it is now showing clear feedback if the user wants.

I was also able to generate some extra simulations based on the examples, so im ticking all the boxes for the issues and review items.

However the paper one i am leaving open for now, since i believe this has not been improved or clarified.
It is not that i dont see use for this tool but i would like to see it better explained what the extra / substantial scholarly effort is for this software.

below ill be a bit more specific and critical about the statements made .

** Paper: **
- [ ] The only argument is that te community needs mesh free simulators. Which this tool indeed brings. But i dont think the meshing is the issue. The solution presented here is twofold i) remove mesh ii) parameterize the environement. The second argument allows for running many easy configurable experiemnts. However, one can also write some code that can easily generate meshes based on the same parameters.
- [ ] The second argument is the GPU implementation. Some simulaters run on CPU on multiple cores. If one does not have a GPU the tool cannot be used. So arguments can also be made for CPU implementation with paralelization. So i dont think dat choices of implementation can be an argument or unmet need. Computation speed could be but only if better supported.
- [ ] With te new improvements i was able to simulate and play a bit. I realized no examples with cells were given so i tried those. However, I cannot specify the diffusivity within the cells. Also from the initial trajectories it seems the cells are overlapping with the fibers.  I gave fiber fractions of .2,.2,.5 and cell fractios of .1, .1 which both are defined in the help as the "the volume of the imaging voxel occupied by..." which adds to more than 100%. Can different populations can occulpy the same volume?
- [ ] The paper also states "highly configurable" which is bit of a stretch since there is only one fiber rotation angle, no curving fibers within voxels (e.g. https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0076626), etc. The actual Emperical volume fraction reported back to me dont match the ones i asked for and the cell fraction of both cells should be equal to the each of fiber 1 and 2. I gues this has to do with fiber and cell pakking, but if i want to simulate a realisticly observed vollume distribution this cannot be done. Yes in potential this framework can be expanded with all these cases but currently it is not.
Im not arguing the tool has no needs, indeed the forward simulation to validated inverse moddeling is a valid one. But the claims and statement of need should be better defined.

```
simDRIFT: Running simDRIFT
simDRIFT: ------------------------------
simDRIFT:  Fiber Setup
simDRIFT: ------------------------------
simDRIFT:  144 fibers of type 0 (R0 = 1.0)
simDRIFT:  144 fibers of type 1 (R1 = 1.0)
simDRIFT:  361 fibers of type 2 (R2 = 1.0)
simDRIFT:  Fiber geometry: Penetrating
simDRIFT: ------------------------------
simDRIFT:  Rotation Matrices
simDRIFT: ------------------------------
simDRIFT:  [[ 1.00,  0.00,  0.00],
simDRIFT:   [ 0.00,  1.00,  0.00],
simDRIFT:   [-0.00,  0.00,  1.00]]

simDRIFT:  [[ 0.71,  0.00,  0.71],
simDRIFT:   [ 0.00,  1.00,  0.00],
simDRIFT:   [-0.71,  0.00,  0.71]]

simDRIFT:  [[-0.71,  0.00,  0.71],
simDRIFT:   [ 0.00,  1.00,  0.00],
simDRIFT:   [-0.71,  0.00, -0.71]]

simDRIFT: ------------------------------
simDRIFT:  Placing Cells...
simDRIFT: ------------------------------
simDRIFT: ------------------------------
simDRIFT:  Cells Setup
simDRIFT: ------------------------------
simDRIFT:  11 cells with radius = 5.0 um
simDRIFT:  1 cells with radius = 10.0 um
simDRIFT: ------------------------------
simDRIFT:  Empirical Volume Fractions
simDRIFT: ------------------------------
simDRIFT:  Fiber 1 Volume: 0.060572
simDRIFT:  Fiber 2 Volume: 0.052422
simDRIFT:  Fiber 3 Volume: 0.125596
simDRIFT:     Cell Volume: 0.040788
simDRIFT: ------------------------------
simDRIFT:  Beginning Simulation...
simDRIFT: ------------------------------
simDRIFT:  Simulation complete!
simDRIFT:  Elapsed time: 268 seconds
```
