### 2023-09-05 8d65b69
Update paper.md
Statement of need + comparison to existing software (disimpy)
