### 2023-09-06 54fc6b8
Update paper.md
better language
