### 2023-09-06 45d3e11
Update paper.md
fixed typos
