### 2023-09-06 aba874e
Update paper.md
redundant figure labels
