+++ joss/paper.md
@@ -30,7 +30,7 @@ Monte Carlo simulations are particularly effective at generating synthetic diffu
 
 ``simDRIFT``'s superior computational performance represents an important advantage relative to other available open-source software. Comparisons between simDRIFT and Disimpy, a mesh based, CUDA enabled DWI forward simulator known to be faster than CAMINO [@Kerkelae2020], on identical image voxel geometries featuring a 5 $\mu m$ radius cell, reveals orders of magnitude improved performance, particularly for large resident spin ensemble sizes. Thus, ``simDRIFT`` is especially useful in a regime where the computational cost of existing software may be cumbersome or even prohibitive. In particular, ``simDRIFT`` is able to perform large scale simulations very quickly, therefore benefiting from favorable convergence properties with respect to the diffusion weighted signal. In this context, using just desktop or even laptop level GPU hardware, ``simDRIFT`` users are able to quickly and easily generate large amounts of synthetic DWI data from a wide variety of voxel geometries.
 
-![Figure 1: Runtime comparison between simDRIFT and Disimpy, another DWI simulator that runs on the GPU. These simulations were performed on a Windows 10 desktop with an Nvidia RTX 3090 GPU.\label{fig:performance}](figs/simDRIFT_vs_Disimpy.png){ width=70% }
+![Runtime comparison between simDRIFT and Disimpy, another DWI simulator that runs on the GPU. These simulations were performed on a Windows 10 desktop with an Nvidia RTX 3090 GPU.\label{fig:performance}](figs/simDRIFT_vs_Disimpy.png){ width=70% }
 
 Therefore, the software encompassed by `simDRIFT` fulfills a presently-unmet need by allowing for mesh-free Monte Carlo simulations of DWI that unify researchers' needs for computational performance and biophysical realism with easy-to-use and configurable open-source software. `simDRIFT` was designed to be used by researchers of all disciplines and focuses who are working with diffusion MRI. Multiple scientific publications which utilize this library are currently in production. The wide customizability, high computational speed, and massively-parallel design will provide avenues for improved model development pipelines and thorough inter-model comparisons, among other potential applications. 
 
@@ -38,7 +38,7 @@ Therefore, the software encompassed by `simDRIFT` fulfills a presently-unmet nee
 The library allows users to construct voxel geometries described by user-defined microstructural and scanning parameters. Specifically, `simDRIFT` simulates the diffusion MRI signal generated from the self-diffusion of water in an isotropic imaging voxel of length $L_{\mathrm{voxel}}$ that contains $n \in [0,4]$ distinct fiber bundles and $m \in [0,2]$ distinct cells types, according to the user's selection of diffusion imaging protocol, diffusion time ($\Delta$), time-step size $\mathrm{d}t$, and desired free water diffusivity $D_{FW}$. Within each simulated voxel, the user also has control over the properties of each fiber/cell type. For each fiber bundle, users define the desired orientation (via the angle $\theta$ formed between the bundle and the $z$ axis), intrinsic diffusivity ($D_{i}$), axonal radius ($R_{i}$), and voxel volume fraction $V_{i}/V_{\mathrm{vox}}$. For each cell type, users similarly define the desired radius $R_{j}$ and voxel volume fraction $V_{j}/V_{\mathrm{vox}}$. Examples of such voxel configurations can be seen in \ref{fig:examples}. 
 
 
-![Figure 2: Example simulated spin trajectories from an imaging voxel featuring two fiber bundles (red, blue) with various orientations ($\theta$ = 0°, 30°, 60°, 90°), along with extra-fiber spins (purple) .\label{fig:examples}](figs/joss_paper_figure.png){ width=70% }
[…]
