### 2023-09-06 #5621 @jacobblum
@djps In the most recent amendment to the simDRIFT manuscript, we have described and presented direct comparisons between simDRIFT and other existing open source software (Disimpy - https://joss.theoj.org/papers/10.21105/joss.02527). We hope that this satisfies this part of the checklist. Thanks so much, and please let us know what additional changes / clarifications, if any, need to be made.

> ## Review checklist for @djps
> * [ ]  **State of the field:** Do the authors describe how this software compares to other commonly-used packages?
