### 2023-09-07 #5621 @jacobblum
> > > @jacobblum - I can't seem to install it when following the instructions:
> > > In windows 11, using conda 23.5.2, (with python 3.11), cuda driver (528.90) and nvcc 11.3.58 but can't install the cuda-toolkit.
> > > It finds lots of conflicts which it can't resolve. What version of conda was this tested on? I had tried different channels also.
> > > I was able to install it using pip though. The tests passed after 28 minutes. I could replicate the quick start examples as well.
> >
> >
> > Hi @djps,
> > We are looking more into what may have caused the errors you experienced when using the installation instructions we provided. At present, we have been unable to replicate the error -- we will get back to you with further clarifying questions should we need more information.
> > In the meantime, we will update the installation instructions to use pip instead of conda to install cudatoolkit since it worked for you (as well as our tests).
> > Thank you! -Kainen
>
> I can't find this in the installation instructions on github or readthedocs.

The updated installation instructions should be in the README and also here: https://simdrift.readthedocs.io/en/latest/installation/index.html
