### 2023-09-11 #5621 @jacobblum
>

> > @djps Can you confirm if you're happy with the last two check-boxes?
> >
> > * Installation: Does installation proceed as outlined in the documentation?
> > * Functionality: Have the functional claims of the software been confirmed?
>
> A one sentence regarding installation using pip would be good, but I am not going to insist on it.

@djps

Thank you for the feedback. We'll add a troubleshooting section to the installation section and mention alternative installation of Numba via pip. I was able to test this out and confirmed it worked on our machines over the weekend.

We've also addressed the following items in the manuscript.

> There is duplication in the captions of the figures, so they read "Figure 1: Figure 1"
>The capitalisation of accronyms is inconsistent in lines 6 and 7.
>On line 39 it should read $\mu$m
