+++ joss/paper.md
@@ -28,7 +28,7 @@ This library, `simDRIFT`, provides for rapid and flexible Monte-Carlo simulation
 # Statement of need
 Monte Carlo simulations are particularly effective at generating synthetic diffusion MRI data from complex, biophysically accurate imaging voxels with known ground-truth microstructural parameters. Consequently, such simulations have proven useful for developing and validating signal processing models [@Chiang2014; @Ye2020]. Existing Monte Carlo simulators, such as CAMINO, Disimpy, Realistic Microstructure Simulator, and others rely on meshes to discretize the computational domain [see, e.g., @Yeh2013, @Ianus2016, @Kerkelae2020, @RafaelPatino2020]. While this approach does allow for the representation of complex and finely detailed microstructural features, given the coarse graining of such features (fiber bending, etc...) observed at experimentally realistic diffusion times and voxel sizes [@Novikov2018], ``simDRIFT`` supplies a reasonable approximation to the relevant self-diffusion processes by parametrizing fibers as narrow cylinders, or "sticks", and cells as isotropic spheres, or "balls" [@Behrens2003]. This way, users may readily simulate voxels induced by multiple, oriented fiber bundles and cells without having to re-discretize the computational domain with complex meshes, which scale in size according to the desired feature resolution and size of the simulated image voxel [@Panagiotaki2010].
 
-``simDRIFT``'s superior computational performance represents an important advantage relative to other available open-source software. Comparisons between simDRIFT and Disimpy, a mesh based, CUDA enabled DWI forward simulator known to be faster than CAMINO [@Kerkelae2020], on identical image voxel geometries featuring a 5 $\mu$ m radius cell, reveals orders of magnitude improved performance, particularly for large resident spin ensemble sizes. Thus, ``simDRIFT`` is especially useful in a regime where the computational cost of existing software may be cumbersome or even prohibitive. In particular, ``simDRIFT`` is able to perform large scale simulations very quickly, therefore benefiting from favorable convergence properties with respect to the diffusion weighted signal. In this context, using just desktop or even laptop level GPU hardware, ``simDRIFT`` users are able to quickly and easily generate large amounts of synthetic DWI data from a wide variety of voxel geometries.
+``simDRIFT``'s superior computational performance represents an important advantage relative to other available open-source software. Comparisons between simDRIFT and Disimpy, a mesh based, CUDA enabled DWI forward simulator known to be faster than CAMINO [@Kerkelae2020], on identical image voxel geometries featuring a 5 $\mu$m radius cell, reveals orders of magnitude improved performance, particularly for large resident spin ensemble sizes. Thus, ``simDRIFT`` is especially useful in a regime where the computational cost of existing software may be cumbersome or even prohibitive. In particular, ``simDRIFT`` is able to perform large scale simulations very quickly, therefore benefiting from favorable convergence properties with respect to the diffusion weighted signal. In this context, using just desktop or even laptop level GPU hardware, ``simDRIFT`` users are able to quickly and easily generate large amounts of synthetic DWI data from a wide variety of voxel geometries.
 
 ![Runtime comparison between simDRIFT and Disimpy, another DWI simulator that runs on the GPU. These simulations were performed on a Windows 10 desktop with an Nvidia RTX 3090 GPU.\label{fig:performance}](figs/simDRIFT_vs_Disimpy.png){ width=70% }
 
