+++ docs/source/troubleshooting/index.rst
@@ -1,0 +1,9 @@
+Troubleshooting
+===================
+
+Installation 
+-------------------
+For certain hardware configurations, we have received notice 
+that installation via ``conda`` triggers certain incompatibilities between
+``numba`` and ``llvmlite``. If you encounter this issue, we reccomend the following installation 
+procedure. 
