### 2023-09-11 0af7e67
Update index.rst
troubleshooting installation
