### 2023-09-11 #5621 @jacobblum
@djps

> ## Review checklist for @djps
> ### Functionality
> * [ ]  **Installation:** Does installation proceed as outlined in the documentation?

Please see our troubleshooting section (https://simdrift.readthedocs.io/en/latest/troubleshooting/index.html) that walks through installation with pip. We have also amended the main installation instructions to include, in the numba step:

>If you encounter any difficulties with this step, please see our [troubleshooting guide](https://simdrift.readthedocs.io/en/latest/troubleshooting/index.html) for an alternative installation procedure.

We hope this satisfies this element of the checklist, and if not please let us know what additional steps you'd like us to take. Thanks so much!
