### 2023-09-16 #5621 @jgostick
Hi @jacobblum, yes you are correct, the review looks complete.  Congrats to you and thank you very much to the reviewers for donating their time to keep the JOSS machine running!

There are some steps we need to take to finalize this.  I will create another checklist of items for you and I to address.
