### 2023-09-16 7d2cb73
Update paper.bib
Updated capitalization
