+++ joss/paper.bib
@@ -1,7 +1,7 @@
 @InProceedings{Lam2015,
   author    = {Lam, Siu Kwan and Pitrou, Antoine and Seibert, Stanley},
   booktitle = {Proceedings of the Second Workshop on the LLVM Compiler Infrastructure in HPC},
-  title     = {Numba: A llvm-based python jit compiler},
+  title     = {Numba: A {LLVM}-based python {JIT} compiler},
   year      = {2015},
   pages     = {1--6},
 }
@@ -39,7 +39,7 @@
 @Article{Kerkelae2020,
   author    = {Leevi Kerkelä and Fabio Nery and Matt Hall and Chris Clark},
   journal   = {Journal of Open Source Software},
-  title     = {Disimpy: A Massively Parallel Monte Carlo Simulator for Generating Diffusion-Weighted {MRI} Data in Python},
+  title     = {Disimpy: A Massively Parallel {M}onte {C}arlo Simulator for Generating Diffusion-Weighted {MRI} Data in Python},
   year      = {2020},
   month     = aug,
   number    = {52},
