### 2023-09-16 4831590
Update paper.bib
more capitalization
