+++ joss/paper.bib
@@ -9,7 +9,7 @@
 @Book{VanRossum2010,
   author    = {VanRossum, Guido and Drake, Fred L.},
   publisher = {Python Software Foundation Amsterdam, Netherlands},
-  title     = {The Python Language Reference},
+  title     = {The {P}ython {L}anguage {R}eference},
   year      = {2010},
 }
 
@@ -39,7 +39,7 @@
 @Article{Kerkelae2020,
   author    = {Leevi Kerkelä and Fabio Nery and Matt Hall and Chris Clark},
   journal   = {Journal of Open Source Software},
-  title     = {Disimpy: A Massively Parallel {M}onte {C}arlo Simulator for Generating Diffusion-Weighted {MRI} Data in Python},
+  title     = {Disimpy: {A} {M}assively {P}arallel {M}onte {C}arlo {S}imulator for {G}enerating {D}iffusion-{W}eighted {MRI} {D}ata in {P}ython},
   year      = {2020},
   month     = aug,
   number    = {52},
@@ -52,7 +52,7 @@
 @Article{Yeh2013,
   author    = {Yeh, Chun-Hung and Schmitt, Benoı̂t and Bihan, Denis Le and Li-Schlittgen, Jing-Rebecca and Lin, Ching-Po and Poupon, Cyril},
   journal   = {{PLoS} {ONE}},
-  title     = {Diffusion Microscopist Simulator: A General Monte Carlo Simulation System for Diffusion Magnetic Resonance Imaging},
+  title     = {Diffusion {M}icroscopist {S}imulator: A {G}eneral {M}onte {C}arlo {S}imulation {S}ystem for {D}iffusion {M}agnetic {R}esonance {I}maging},
   year      = {2013},
   number    = {10},
   pages     = {e76626},
@@ -65,7 +65,7 @@
 @Article{RafaelPatino2020,
   author    = {Rafael-Patino, Jonathan and Romascano, David and Ramirez-Manzanares, Alonso and Canales-Rodrı́guez, Erick Jorge and Girard, Gabriel and Thiran, Jean-Philippe},
   journal   = {Frontiers in Neuroinformatics},
-  title     = {Robust Monte-Carlo Simulations in Diffusion-{MRI}: Effect of the Substrate Complexity and Parameter Choice on the Reproducibility of Results},
+  title     = {Robust {M}onte-{C}arlo {S}imulations in {D}iffusion-{MRI}: {E}ffect of the {S}ubstrate {C}omplexity and {P}arameter {C}hoice on the {R}eproducibility of {R}esults},
   year      = {2020},
   month     = mar,
   volume    = {14},
@@ -77,7 +77,7 @@
   author    = {Panagiotaki, Eleftheria and Hall, Matt G. and Zhang, Hui and Siow, Bernard and Lythgoe, Mark F. and Alexander, Daniel C.},
   booktitle = {Medical Image Computing and Computer-Assisted Intervention {\textendash} {MICCAI} 2010},
   publisher = {Springer Berlin Heidelberg},
-  title     = {High-Fidelity Meshes from Tissue Samples for Diffusion {MRI} Simulations},
+  title     = {High-{F}idelity {M}eshes from {T}issue {S}amples for {D}iffusion {MRI} {S}imulations},
   year      = {2010},
   pages     = {404--411},
   doi       = {10.1007/978-3-642-15745-5_50},
