### 2023-09-16 #5621 @jacobblum
@jgostick

Thank you for the message. Everything should be set for the author tasks.

I've added a new release version `1.0.13` with a Git tag of `v1.0.13`
and added the repository to Zenodo. The DOI is [10.5281/zenodo.8351982](https://doi.org/10.5281/zenodo.8351982)

Thanks!
