### 2023-09-16 5c342eb
Update paper.md
Timestep notation
