### 2023-09-16 #5621 @jgostick
Hi @jacobblum, one the items on my checklist is to check the formatting of the references.  In some of your references you use sentence case, while in other you use capital case for the paper title.  Can you change these to sentence case?

 I realize this is very minor, and maybe our pdf builder will one day fix this automatically.
