### 2023-09-17 3134822
Update paper.bib
sentence case
