+++ joss/paper.bib
@@ -1,14 +1,14 @@
 @InProceedings{Lam2015,
   author    = {Lam, Siu Kwan and Pitrou, Antoine and Seibert, Stanley},
-  booktitle = {Proceedings of the Second Workshop on the LLVM Compiler Infrastructure in HPC},
-  title     = {Numba: A {LLVM}-based python {JIT} compiler},
+  booktitle = {Proceedings of the second workshop on the llvm compiler infrastructure in hpc},
+  title     = {Numba: A llvm-based python jit compiler},
   year      = {2015},
   pages     = {1--6},
 }
 
 @Book{VanRossum2010,
   author    = {VanRossum, Guido and Drake, Fred L.},
-  publisher = {Python Software Foundation Amsterdam, Netherlands},
+  publisher = {Python software foundation Amsterdam, Netherlands},
   title     = {The {P}ython {L}anguage {R}eference},
   year      = {2010},
 }
@@ -28,7 +28,7 @@
 @Article{Ye2020,
   author    = {Ye, Zezhong and Gary, Sam E. and Sun, Peng and Mustafi, Sourajit Mitra and Glenn, George Russell and Yeh, Fang-Cheng and Merisaari, Harri and Song, Chunyu and Yang, Ruimeng and Huang, Guo-Shu and Kao, Hung-Wen and Lin, Chien-Yuan and Wu, Yu-Chien and Jensen, Jens H. and Song, Sheng-Kwei},
   journal   = {{NMR} in Biomedicine},
-  title     = {The impact of edema and fiber crossing on diffusion {MRI} metrics assessed in an ex vivo nerve phantom: {M}ulti-tensor model {vs.} diffusion orientation distribution function},
+  title     = {The impact of edema and fiber crossing on diffusion mri metrics assessed in an ex vivo nerve phantom: multi-tensor model {vs.} diffusion orientation distribution function},
   year      = {2020},
   number    = {1},
   volume    = {34},
@@ -39,7 +39,7 @@
 @Article{Kerkelae2020,
   author    = {Leevi Kerkelä and Fabio Nery and Matt Hall and Chris Clark},
   journal   = {Journal of Open Source Software},
-  title     = {Disimpy: {A} {M}assively {P}arallel {M}onte {C}arlo {S}imulator for {G}enerating {D}iffusion-{W}eighted {MRI} {D}ata in {P}ython},
+  title     = {Disimpy: a massively parallel monte carlo simulator for generating diffusion-weighted mri data in python},
   year      = {2020},
   month     = aug,
   number    = {52},
@@ -52,7 +52,7 @@
 @Article{Yeh2013,
   author    = {Yeh, Chun-Hung and Schmitt, Benoı̂t and Bihan, Denis Le and Li-Schlittgen, Jing-Rebecca and Lin, Ching-Po and Poupon, Cyril},
   journal   = {{PLoS} {ONE}},
-  title     = {Diffusion {M}icroscopist {S}imulator: A {G}eneral {M}onte {C}arlo {S}imulation {S}ystem for {D}iffusion {M}agnetic {R}esonance {I}maging},
+  title     = {Diffusion microscopist simulator: A general monte carlo simulation system for diffusion magnetic resonance imaging},
   year      = {2013},
   number    = {10},
   pages     = {e76626},
@@ -65,7 +65,7 @@
 @Article{RafaelPatino2020,
   author    = {Rafael-Patino, Jonathan and Romascano, David and Ramirez-Manzanares, Alonso and Canales-Rodrı́guez, Erick Jorge and Girard, Gabriel and Thiran, Jean-Philippe},
   journal   = {Frontiers in Neuroinformatics},
[…]
