### 2023-09-17 be233de
Update paper.bib
sentence case pt. 2
