+++ joss/paper.bib
@@ -1,6 +1,6 @@
 @InProceedings{Lam2015,
   author    = {Lam, Siu Kwan and Pitrou, Antoine and Seibert, Stanley},
-  booktitle = {Proceedings of the \lowercase{s}econd \lowercase{w}orkshop on the \lowercase{l}lvm \lowercase{c}ompiler \lowercase{i}nfrastructure in \lowercase{h}pc},
+  booktitle = {Proceedings of the second workshop on the llvm compiler infrastructure in hpc},
   title     = {Numba: A llvm-based python jit compiler},
   year      = {2015},
   pages     = {1--6},
@@ -28,7 +28,7 @@
 @Article{Ye2020,
   author    = {Ye, Zezhong and Gary, Sam E. and Sun, Peng and Mustafi, Sourajit Mitra and Glenn, George Russell and Yeh, Fang-Cheng and Merisaari, Harri and Song, Chunyu and Yang, Ruimeng and Huang, Guo-Shu and Kao, Hung-Wen and Lin, Chien-Yuan and Wu, Yu-Chien and Jensen, Jens H. and Song, Sheng-Kwei},
   journal   = {{NMR} in Biomedicine},
-  title     = {The impact of edema and fiber crossing on diffusion mri metrics assessed in an ex vivo nerve phantom: multi-tensor model vs. diffusion orientation distribution function},
+  title     = {The impact of edema and fiber crossing on diffusion mri metrics assessed in an ex vivo nerve phantom test: multi-tensor model vs. diffusion orientation distribution function},
   year      = {2020},
   number    = {1},
   volume    = {34},
