### 2023-09-17 29b9882
Update paper.bib
final capitalization
