+++ joss/paper.bib
@@ -1,6 +1,6 @@
 @InProceedings{Lam2015,
   author    = {Lam, Siu Kwan and Pitrou, Antoine and Seibert, Stanley},
-  booktitle = {Proceedings of the second workshop on the llvm compiler infrastructure in hpc},
+  booktitle = {Proceedings of the {s}econd {w}orkshop on the {l}lvm {c}ompiler {i}nfrastructure in {h}pc},
   title     = {Numba: A llvm-based python jit compiler},
   year      = {2015},
   pages     = {1--6},
