### 2023-09-17 #5621 @jacobblum
Hi @jgostick, thank you for catching that! The reference section should be consistently in sentence case now - compiling the draft locally it looks right. Please let me know if the formatting is okay, and if not what needs to be amended. Thanks so much!
