### 2023-09-17 830d1b6
Update paper.md
Citation fix
