### 2023-09-17 8d2920d
Update paper.md
fixed citation in statement of need
