### 2023-09-17 c849401
minor changes in paper
also trying to fix the second set of references in the second paragraph, which didn't build correctly in the current draft
