### 2023-09-17 a140663
Update paper.bib - partial
