+++ joss/paper.bib
@@ -1,7 +1,7 @@
 @InProceedings{Lam2015,
   author    = {Lam, Siu Kwan and Pitrou, Antoine and Seibert, Stanley},
-  booktitle = {Proceedings of the {s}econd {w}orkshop on the {l}lvm {c}ompiler {i}nfrastructure in {h}pc},
-  title     = {Numba: A llvm-based python jit compiler},
+  booktitle = {Proceedings of the second workshop on the {LLVM} compiler infrastructure in {HPC}},
+  title     = {Numba: A {LLVM}-based {P}ython {JIT} compiler},
   year      = {2015},
   pages     = {1--6},
 }
@@ -39,7 +39,7 @@
 @Article{Kerkelae2020,
   author    = {Leevi Kerkelä and Fabio Nery and Matt Hall and Chris Clark},
   journal   = {Journal of Open Source Software},
-  title     = {Disimpy: a massively parallel monte carlo simulator for generating diffusion-weighted mri data in python},
+  title     = {Disimpy: a massively parallel {M}onte {C}arlo simulator for generating diffusion-weighted {MRI} data in {P}ython},
   year      = {2020},
   month     = aug,
   number    = {52},
@@ -52,7 +52,7 @@
 @Article{Yeh2013,
   author    = {Yeh, Chun-Hung and Schmitt, Benoı̂t and Bihan, Denis Le and Li-Schlittgen, Jing-Rebecca and Lin, Ching-Po and Poupon, Cyril},
   journal   = {{PLoS} {ONE}},
-  title     = {Diffusion microscopist simulator: A general monte carlo simulation system for diffusion magnetic resonance imaging},
+  title     = {Diffusion microscopist simulator: A general {M}onte {C}arlo simulation system for diffusion magnetic resonance imaging},
   year      = {2013},
   number    = {10},
   pages     = {e76626},
@@ -86,7 +86,7 @@
 @Article{Ianus2016,
   author    = {Andrada Ianuş and Ivana Drobnjak and Daniel C. Alexander},
   journal   = {{NMR} in Biomedicine},
-  title     = {Model-based estimation of microscopic anisotropy using diffusion mri: a simulation study},
+  title     = {Model-based estimation of microscopic anisotropy using diffusion {MRI}: a simulation study},
   year      = {2016},
   month     = mar,
   number    = {5},
