### 2023-09-17 3bfa9ee
Update paper.bib - final
