+++ joss/paper.bib
@@ -28,7 +28,7 @@
 @Article{Ye2020,
   author    = {Ye, Zezhong and Gary, Sam E. and Sun, Peng and Mustafi, Sourajit Mitra and Glenn, George Russell and Yeh, Fang-Cheng and Merisaari, Harri and Song, Chunyu and Yang, Ruimeng and Huang, Guo-Shu and Kao, Hung-Wen and Lin, Chien-Yuan and Wu, Yu-Chien and Jensen, Jens H. and Song, Sheng-Kwei},
   journal   = {{NMR} in Biomedicine},
-  title     = {The impact of edema and fiber crossing on diffusion mri metrics assessed in an ex vivo nerve phantom: multi-tensor {m}odel vs. {d}iffusion orientation distribution function},
+  title     = {The impact of edema and fiber crossing on diffusion {MRI} metrics assessed in an ex vivo nerve phantom: multi-tensor {m}odel vs. {d}iffusion orientation distribution function},
   year      = {2020},
   number    = {1},
   volume    = {34},
@@ -65,7 +65,7 @@
 @Article{RafaelPatino2020,
   author    = {Rafael-Patino, Jonathan and Romascano, David and Ramirez-Manzanares, Alonso and Canales-Rodrı́guez, Erick Jorge and Girard, Gabriel and Thiran, Jean-Philippe},
   journal   = {Frontiers in Neuroinformatics},
-  title     = {Robust monte-carlo simulations in diffusion-mri: {E}ffect of the substrate complexity and parameter choice on the reproducibility of results},
+  title     = {Robust {M}onte-{C}arlo simulations in diffusion-{MRI}: Effect of the substrate complexity and parameter choice on the reproducibility of results},
   year      = {2020},
   month     = mar,
   volume    = {14},
