### 2023-09-17 #5621 @danielskatz
Also, I've found a few things to suggest in the paper and bib - in https://github.com/jacobblum/simDRIFT/pull/15  Please merge this, or let me know what you disagree with, then we can proceed.
