### 2023-09-17 #5621 @danielskatz
Congratulations to @jacobblum (Jacob Blum) and co-author on your publication!!

JOSS is dependent on volunteers, and we couldn't do this without you
